Add .env and .gitignore files
The .gitignore file tells Git to ignore certain files in the repository
directory, and not show them when running `git status`. This helps
against accidentally committing files that aren't part of the code.

In particular, we want Git to ignore .env so that we don't accidentally
commit secrets such as passwords to the repository.  We also ignore
files in the __pycache__/ directory, which Python uses to store
temporary compiled code, and the .idea/ directory, which is generated by
the PyCharm IDE.

However, to help new deployments, we also include a .env file that has
all the names of the environment variables that will need to be
provided.